backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
openai
python-dotenv
requests
beautifulsoup4
